Send analysis only if changed or first check

+++ udata_hydra/analysis.py
@@ -47,16 +47,15 @@ async def download_resource(url: str, headers: dict) -> BinaryIO:
     return tmp_file
 
 
-# TODO: we're sending analysis info to udata every time a resource is crawled
-# although we only send "meta" check info only when they change
-# maybe introduce conditionnal webhook trigger here too
-async def process_resource(check_id: int) -> None:
+async def process_resource(check_id: int, is_first_check: bool) -> None:
     """
     Perform analysis on the resource designated by check_id
     - change analysis
-    - size
-    - mime_type
-    # FIXME: saving to minio and parsing CSV not called for now
+    - size (optionnal)
+    - mime_type (optionnal)
+    - checksum (optionnal)
+
+    Will call udata if first check or changes found, and update check with optionnal infos
     """
     check = await get_check(check_id)
     if not check:
@@ -75,46 +74,43 @@ async def process_resource(check_id: int) -> None:
     # if not, let's see if we can infer a modifification date from headers
     change_analysis = change_analysis or await detect_resource_change_from_headers(url) or {}
 
-    tmp_file = None
-    sha1 = None
-    error = None
-    # TODO: those can be infered from header, we could skip download totally
-    mime_type = None
-    filesize = None
+    # if not, let's download the file to get some hints and other infos
+    dl_analysis = {}
+    if not change_analysis:
+        tmp_file = None
+        try:
+            tmp_file = await download_resource(url, headers)
+        except IOError:
+            dl_analysis["analysis:error"] = "File too large to download"
+        else:
+            # Get file size
+            dl_analysis["analysis:filesize"] = os.path.getsize(tmp_file.name)
+            # Get checksum
+            dl_analysis["analysis:checksum"] = compute_checksum_from_file(tmp_file.name)
+            # Check if checksum has been modified if we don't have other hints
+            change_analysis = (
+                await detect_resource_change_from_checksum(resource_id, dl_analysis["analysis:checksum"])
+                or {}
+            )
+            # TODO: this never seems to output text/csv, maybe override it later
+            dl_analysis["analysis:mime-type"] = magic.from_file(tmp_file.name, mime=True)
+        finally:
+            if tmp_file:
+                os.remove(tmp_file.name)
+            await update_check(check_id, {
+                "checksum": dl_analysis.get("analysis:checksum"),
+                "analysis_error": dl_analysis.get("analysis:error"),
+                "filesize": dl_analysis.get("analysis:filesize"),
+                "mime_type": dl_analysis.get("analysis:mime-type"),
+            })
 
-    try:
-        tmp_file = await download_resource(url, headers)
-    except IOError:
-        error = "File too large to download"
-    else:
-        # Get file size
-        filesize = os.path.getsize(tmp_file.name)
-        # Get checksum
-        sha1 = compute_checksum_from_file(tmp_file.name)
-        # Check if checksum has been modified if we don't have other hints
-        change_analysis = change_analysis or await detect_resource_change_from_checksum(resource_id, sha1) or {}
-        # FIXME: this never seems to output text/csv, maybe override it later
-        mime_type = magic.from_file(tmp_file.name, mime=True)
-    finally:
-        if tmp_file:
-            os.remove(tmp_file.name)
+    if change_analysis or (is_first_check and dl_analysis):
+        change_analysis.update(dl_analysis)
         await send(
             dataset_id=dataset_id,
             resource_id=resource_id,
-            document={
-                "analysis:error": error,
-                "analysis:filesize": filesize,
-                "analysis:mime-type": mime_type,
-                "analysis:checksum": sha1,
-                **change_analysis,
-            }
+            document=change_analysis
         )
-        await update_check(check_id, {
-            "checksum": sha1,
-            "analysis_error": error,
-            "filesize": filesize,
-            "mime_type": mime_type
-        })
 
 
 async def detect_resource_change_from_checksum(resource_id, new_checksum):
